requirements.txt
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cycler==0.11.0
ekphrasis==0.5.1
emoji==1.6.1
filelock==3.3.2
ftfy==6.0.3
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.11.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.1.2
hyperframe==5.2.0
idna==2.10
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.4.3
nltk==3.6.5
numpy==1.21.4
packaging==21.2
pandas==1.3.4
Pillow==8.4.0
pyenchant==3.2.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
rfc3986==1.5.0
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
sklearn==0.0
sniffio==1.2.0
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.10.0
torchtext==0.11.0
tqdm==4.62.3
transformers==4.12.3
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
wcwidth==0.2.5
